add settings
